requirements.txt
streamlit>=1.28.0
langchain-google-genai>=0.0.6
yfinance>=0.2.18
duckduckgo-search>=3.9.6
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0